config: configurando caminho global dos arquivos e refatorando estrutura de pastas
