+++ src/features/sobre-nos/pages/index.tsx
@@ -1,0 +1,5 @@
+export default function SobreNos(){
+    return<main>
+        <h1>OI</h1>
+    </main>
+}
