--- src/pagas/sobreNos/index.tsx
@@ -1,5 +1,0 @@
-export default function SobreNos(){
-    return<main>
-        <h1>OI</h1>
-    </main>
-}
